==> src/02_autoencoderBoruta_s1.Rmd <==
---
title: "02_autoencoder.Rmd"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(here)
library(tidyverse)
library(Boruta)
library(caret)
library(h2o)
source(here("/src/utils.R"))

h2o.init()

seed <- 17
```

load data
```{r}
s1_rna_features <- readRDS(file = here("munge", "s1_rna_features.rds"))
s1_features_with_outcome <- readRDS(file = here("munge", "s1_features_with_outcome.rds"))
s1_features_with_outcome_phenotype<-readRDS(file = here("munge", "s1_features_with_outcome_phenotype.rds"))
```


```{r}
features <- colnames(s1_rna_features)
train.data.h2o <- as.h2o(s1_rna_features)
anomaly_model <- h2o.deeplearning(x = features, training_frame = train.data.h2o, 
                                  activation = "Tanh",   autoencoder = TRUE,   hidden = 100,  
                                  epochs =100, variable_importances=TRUE, seed=seed, standardize=TRUE)
rna.import.matrix <- anomaly_model@model$variable_importances

```

save important features
```{r}
impRate <- 0.1
import.feature <- rna.import.matrix$variable[1:(nrow(rna.import.matrix)*impRate)]
variable.import <- anomaly_model@model$variable_importances
write.table(variable.import,file = here("/munge/rna_autoVarImport.txt") ,row.names = F,col.names = T,sep = "\t",quote = F)

```

only keep the important features
```{r}

s1_features_with_outcome_phenotype <- s1_features_with_outcome_phenotype[, which(colnames(s1_features_with_outcome_phenotype) %in% c("PATIENTID", "SURVIVAL_STATUS", "SEX", "CANCER_TYPEASTROCYTOMA", "CANCER_TYPEGBM", "CANCER_TYPEMIXED",              "CANCER_TYPEOLIGODENDROGLIOMA", import.feature))]

```


split train and test data
```{r}
s1_features_with_outcome_phenotype <- s1_features_with_outcome_phenotype %>% 
                            mutate (SURVIVAL_STATUS = as.factor (SURVIVAL_STATUS)) %>%
                            rename (class = SURVIVAL_STATUS)

set.seed(seed)
a <- createDataPartition(s1_features_with_outcome_phenotype$class, 
                         p = 0.8, list=FALSE)

s1_train.data_with_outcome_phenotype <- s1_features_with_outcome_phenotype[a,]
s1_test.data_with_outcome_phenotype <- s1_features_with_outcome_phenotype[-a,]


saveRDS(s1_train.data_with_outcome_phenotype, file = here("munge", "s1_train.data_with_outcome_phenotype.rds"))
saveRDS(s1_test.data_with_outcome_phenotype, file = here("munge", "s1_test.data_with_outcome_phenotype.rds"))


```


```{r}

train_data_boruta <- s1_train.data_with_outcome_phenotype %>% select (-PATIENTID, -SEX, -CANCER_TYPEASTROCYTOMA, -CANCER_TYPEGBM, -CANCER_TYPEMIXED, -CANCER_TYPEOLIGODENDROGLIOMA)

predictors.ensemble <- unbalancedFeatureSelectionFromBoruta(train_data_boruta, seed, 0.1)

```



save selected features
```{r}
predictors.all.result <- unique(c(unlist(predictors.ensemble), "SEX", "CANCER_TYPEASTROCYTOMA", "CANCER_TYPEGBM", "CANCER_TYPEMIXED", "CANCER_TYPEOLIGODENDROGLIOMA"))
predictors.overlap.result <- unique(c( unlist(predictors.ensemble)[which(table(unlist(predictors.ensemble))>1)], "SEX", "CANCER_TYPEASTROCYTOMA", "CANCER_TYPEGBM", "CANCER_TYPEMIXED", "CANCER_TYPEOLIGODENDROGLIOMA"))
                                     

write.table(predictors.all.result, file = here("/munge/rna_borutaAllFeatures.txt") ,row.names = F,col.names = F,sep = "\t",quote = F)
write.table(predictors.overlap.result, file = here("/munge/rna_borutaOverlapFeatures.txt") ,row.names = F,col.names = F,sep = "\t",quote = F)

```
