==> src/05_autoencoderBoruta_s2.Rmd <==
---
title: "02_autoencoder.Rmd"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(here)
library(tidyverse)
library(Boruta)
library(caret)
library(h2o)
source(here("/src/utils.R"))

h2o.init()

seed <- 17
```

load data
```{r}
s2_CN_features <- readRDS(file = here("munge", "s2_CN_features.rds"))
s2_features_with_outcome <- readRDS(file = here("munge", "s2_features_with_outcome.rds"))
s2_features_with_outcome_phenotype<-readRDS(file = here("munge", "s2_features_with_outcome_phenotype.rds"))
```


```{r}
cn_features <- colnames(s2_CN_features)
s2_CN_features.h2o <- as.h2o(s2_CN_features)
anomaly_model_CN <- h2o.deeplearning(x = cn_features, training_frame = s2_CN_features.h2o, 
                                  activation = "Tanh",   autoencoder = TRUE,   hidden = 100,  
                                  epochs =100, variable_importances=TRUE, seed=seed, standardize=TRUE)
CN.import.matrix <- anomaly_model_CN@model$variable_importances

```

save important features
```{r}
impRate <- 0.1
import.feature_CN <- CN.import.matrix$variable[1:(nrow(CN.import.matrix)*impRate)]
variable.import_CN <- anomaly_model_CN@model$variable_importances
write.table(variable.import_CN, file = here("/munge/cn_autoVarImport.txt"), 
            row.names = F, col.names = T, sep = "\t", quote = F)

```

only keep the important features
```{r}

s2_features_with_outcome_phenotype <- s2_features_with_outcome_phenotype[, which(colnames(s2_features_with_outcome_phenotype) %in% c("PATIENTID", "SURVIVAL_STATUS", "SEX", "CANCER_TYPEASTROCYTOMA", "CANCER_TYPEGBM", "CANCER_TYPEMIXED",              "CANCER_TYPEOLIGODENDROGLIOMA",     import.feature_CN))]

```


split train and test data
```{r}
s2_features_with_outcome_phenotype <- s2_features_with_outcome_phenotype %>% 
                            mutate (SURVIVAL_STATUS = as.factor (SURVIVAL_STATUS)) %>%
                            rename (class = SURVIVAL_STATUS)

a <- createDataPartition(s2_features_with_outcome_phenotype$class, 
                         p = 0.8, list=FALSE)
s2_train.data_with_outcome_phenotype <- s2_features_with_outcome_phenotype[a,]
s2_test.data_with_outcome_phenotype <- s2_features_with_outcome_phenotype[-a,]

saveRDS(s2_train.data_with_outcome_phenotype, file = here("munge", "s2_train.data_with_outcome_phenotype.rds"))
saveRDS(s2_test.data_with_outcome_phenotype, file = here("munge", "s2_test.data_with_outcome_phenotype.rds"))
```


```{r}

train_data_boruta <- s2_train.data_with_outcome_phenotype %>% select (-PATIENTID, -SEX, -CANCER_TYPEASTROCYTOMA, -CANCER_TYPEGBM, -CANCER_TYPEMIXED, -CANCER_TYPEOLIGODENDROGLIOMA)

predictors.ensemble_CN <- unbalancedFeatureSelectionFromBoruta(train_data_boruta, seed, 0.1)

```


over sample the minority class when doing Boruta
```{r, eval=FALSE}
    flag <- which.max(c(length(which(train.data$class == 0)),length(which(train.data$class == 1))))
    
    minority.class.data <- train.data[which(train.data[,ncol(train.data)] != (flag-1)),]
    majority.class.data <- train.data[which(train.data[,ncol(train.data)] == (flag-1)),]
    class.times <- floor(nrow(majority.class.data)/nrow(minority.class.data))
    set.seed(seed)
    folds.ensemble <- createFolds(y=majority.class.data$class,k = class.times)
    ensemble.i <- 1
    predictors.ensemble <- list()
    
    for (ensemble.i in 1:class.times) {
      print(ensemble.i)
      set.seed(seed)
      ensemble.i.data <- rbind(minority.class.data,majority.class.data[folds.ensemble[[ensemble.i]],])
      boruta.train <- Boruta(class~.,data=ensemble.i.data,pValue=0.01,maxRuns = 200,mcAdj=FALSE)
      feature.boruta <- gsub("\`","",getSelectedAttributes(boruta.train, withTentative = T))
      predictors.ensemble[[ensemble.i]] <- feature.boruta
      t <- 1
      while ((length(feature.boruta) <= 1)&t<=20) {
        print("few predictors left, run again")
        boruta.train <- Boruta(class~.,data=ensemble.i.data,pValue=0.01,maxRuns = 200,mcAdj=FALSE)
        feature.boruta <- gsub("\`","",getSelectedAttributes(boruta.train, withTentative = T))
        predictors.ensemble[[ensemble.i]] <- feature.boruta
        t <- t+1
      }
    }
    
    
    
```

save selected features
```{r}
predictors.all.result_CN <- unique(c(unlist(predictors.ensemble_CN), "SEX", "CANCER_TYPEASTROCYTOMA", "CANCER_TYPEGBM", "CANCER_TYPEMIXED", "CANCER_TYPEOLIGODENDROGLIOMA"))
predictors.overlap.result_CN <- unique(c( unlist(predictors.ensemble_CN)[which(table(unlist(predictors.ensemble_CN))>1)], "SEX", "CANCER_TYPEASTROCYTOMA", "CANCER_TYPEGBM", "CANCER_TYPEMIXED", "CANCER_TYPEOLIGODENDROGLIOMA"))
                                     

write.table(predictors.all.result_CN, file = here("/munge/cn_borutaAllFeatures.txt") ,row.names = F,col.names = F,sep = "\t",quote = F)
write.table(predictors.overlap.result_CN, file = here("/munge/cn_borutaOverlapFeatures.txt") ,row.names = F,col.names = F,sep = "\t",quote = F)

```
